# reference/openapi.yaml
openapi: 3.0.3
info:
  title: MAGNUS API
  description: |
    MAGNUS game-master API for managing campaigns, terminals, characters, and players.

    **Authentication:** JWT Bearer tokens (HS256, 24h expiry). Obtain via `POST /auth/login`.

    **Access control:**
    - Anonymous — public + active resources only
    - Player — public + active + assigned campaigns, own characters only
    - Admin — all resources

    **404 policy:** The API returns 404 for both "not found" and "access denied" to avoid leaking existence.
  version: 1.0.0
servers:
  - url: /
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ObjectId:
      type: string
      pattern: '^[a-f0-9]{24}$'
      example: '65f1a2b3c4d5e6f7a8b9c0d1'
    NanoId:
      type: string
      minLength: 8
      maxLength: 8
      example: 'aB3dE8gH'
...
    Species:
      type: string
      enum: [human, ghoul, super_mutant, robot]
...
    SkillLevel:
      type: string
      enum: [competent, expert, master]
    ConditionSeverity:
      type: string
      enum: [minor, major]
    TagType:
      type: string
      enum: [core, extra]
    PaTrackedBy:
      type: string
      enum: [agility, endurance]
...
    SpecialSection:
      type: object
      properties:
        strength:
          type: integer
          minimum: 1
          maximum: 5
        perception:
          type: integer
          minimum: 1
          maximum: 5
        endurance:
          type: integer
          minimum: 1
          maximum: 5
        charisma:
          type: integer
          minimum: 1
          maximum: 5
        intelligence:
          type: integer
          minimum: 1
          maximum: 5
        agility:
          type: integer
          minimum: 1
          maximum: 5
        luck:
          type: integer
          minimum: 1
          maximum: 5
    CharacterSkill:
      type: object
      required: [id, level]
      properties:
        id:
          type: string
          description: Catalog slug (e.g. "lockpick") — caller-supplied
        level:
          $ref: '#/components/schemas/SkillLevel'
    ActionPoints:
      type: object
      properties:
        paMax:
          type: integer
        paCurrent:
          type: integer
        paTrackedBy:
          $ref: '#/components/schemas/PaTrackedBy'
    Condition:
      type: object
      required: [id, name, severity]
      properties:
        id:
          $ref: '#/components/schemas/NanoId'
        name:
          type: string
        severity:
          $ref: '#/components/schemas/ConditionSeverity'
        description:
          type: string
    StatusSection:
      type: object
      properties:
        positiveConditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        negativeConditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        criticalState:
          type: boolean
    CharacterPerk:
      type: object
      required: [id, name]
      properties:
        id:
          $ref: '#/components/schemas/NanoId'
        name:
          type: string
        description:
          type: string
        icon:
          type: string
    ResourcesSection:
      type: object
      properties:
        caps:
          type: integer
        bobbleheads:
          type: integer
          description: Admin-write only
        scraps:
          type: integer
    Tag:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/TagType'
        damaged:
          type: boolean
    WeaponItem:
      type: object
      required: [id, name]
      properties:
        id:
          $ref: '#/components/schemas/NanoId'
        name:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        broken:
          type: boolean
    ConsumableItem:
      type: object
      required: [id, name]
      properties:
        id:
          $ref: '#/components/schemas/NanoId'
        name:
          type: string
        description:
          type: string
        quantity:
          type: integer
          default: 0
    InventorySection:
      type: object
      properties:
        weapons:
          type: array
          items:
            $ref: '#/components/schemas/WeaponItem'
        equip:
          type: array
          items:
            $ref: '#/components/schemas/WeaponItem'
        consumables:
          type: array
          items:
            $ref: '#/components/schemas/ConsumableItem'
        other:
          type: array
          items:
            $ref: '#/components/schemas/ConsumableItem'
    CharacterResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        campaignId:
          $ref: '#/components/schemas/ObjectId'
        userId:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        species:
          $ref: '#/components/schemas/Species'
        special:
          $ref: '#/components/schemas/SpecialSection'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/CharacterSkill'
        actionPoints:
          $ref: '#/components/schemas/ActionPoints'
        status:
          $ref: '#/components/schemas/StatusSection'
        perks:
          type: array
          items:
            $ref: '#/components/schemas/CharacterPerk'
        resources:
          $ref: '#/components/schemas/ResourcesSection'
        inventory:
          $ref: '#/components/schemas/InventorySection'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCharacterDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        species:
          $ref: '#/components/schemas/Species'
        userId:
          $ref: '#/components/schemas/ObjectId'
          description: Required when caller is admin; ignored for players

    # ─── Configuration ───────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or caller cannot access it
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NoContent:
      description: Success with no body

# ═══════════════════════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /campaigns/{campaignId}/characters:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'

    get:
      tags: [Characters]
      summary: List characters in campaign (actor-scoped)
      description: Players see only their own characters; admins see all.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Character list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Characters]
      summary: Create character in campaign
      description: |
        Players create characters for themselves (userId ignored).
        Admins must supply userId to assign the character to a player.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterDto'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'

    get:
      tags: [Characters]
      summary: Get character detail
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Character detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Characters]
      summary: Full character update
      description: |
        Admin-only fields (`special`, `skills`, `perks`) are silently dropped when sent by a player.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterResponse'
      responses:
        '200':
          description: Updated character
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Characters]
      summary: Soft-delete character
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/special:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch SPECIAL stats (admin-write only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecialSection'
      responses:
        '200':
          description: Updated SPECIAL section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecialSection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/skills:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch skills (admin-write only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CharacterSkill'
      responses:
        '200':
          description: Updated skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CharacterSkill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/perks:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch perks (admin-write only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CharacterPerk'
      responses:
        '200':
          description: Updated perks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CharacterPerk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/status:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch status (conditions + critical state)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusSection'
      responses:
        '200':
          description: Updated status section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusSection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/action-points:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch action points
      description: Players may only update `paCurrent`. `paMax` and `paTrackedBy` are admin-only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionPoints'
      responses:
        '200':
          description: Updated action points section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionPoints'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/inventory:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch inventory items
      description: |
        Omit `id` to create a new item (server mints a nanoid). Provide `id` to update or delete.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventorySection'
      responses:
        '200':
          description: Updated inventory section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventorySection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{campaignId}/characters/{characterId}/resources:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: characterId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    patch:
      tags: [Characters]
      summary: Patch resource counters
      description: Players may update `caps` and `scraps`. `bobbleheads` is admin-only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourcesSection'
      responses:
        '200':
          description: Updated resources section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesSection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'